refactor(api): move logic from route to controller
